Changed the way injuries and cards work
I had a little issue too with the toString method in reason in
combination with the XML Handler. I changed it so the toString method
returns things in all caps now...  Also Fixed a spelling error so if you
get exceptions please rename all occurances of "HARMSTRING" in your
teams.xml file with "HAMSTRING". Or just start a new game :)